@@ -1,11 +1,11 @@
 <script lang="ts">
+  import "../app.postcss"
+  import { page } from "$app/stores"
   import Nav from "../components/Nav.svelte"
   import Header from "../components/Header.svelte"
   import Footer from "../components/Footer.svelte"
   import PersistentModal from "../components/Modal/PersistentModal.svelte"
   import { showDonateModal } from "../data/stores"
-
-  export let segment: string
 </script>
 
 <style>
@@ -27,11 +27,12 @@
     id="ELEOForm"
     style="width:100%;max-width:900px;min-height:1000px;border-width:0px;border-style:none;"
     scrolling="no"
-    src="https://www.eleoonline.net/Pages/WebForms/Mobile/ShowFormMobile.aspx?id=f790c257-b67f-4508-9ff9-0fc3a16f04d2&linkto=670"></iframe>
+    src="https://www.eleoonline.net/Pages/WebForms/Mobile/ShowFormMobile.aspx?id=f790c257-b67f-4508-9ff9-0fc3a16f04d2&linkto=670"
+  ></iframe>
 </PersistentModal>
 
 <Header />
-<Nav segment="{segment}" />
+<Nav activePath="{$page.path}" />
 
 <main class="relative m-auto py-8 px-2 sm:px-8">
   <slot />
