@@ -1,12 +1,20 @@
 <script lang="ts" context="module">
-  export async function preload({ params, query }) {
-    const res = await this.fetch(`data/productions/${params.slug}.json`)
+  import type { Load } from "@sveltejs/kit"
+  export const load: Load = async (obj) => {
+    const res = await obj.fetch(
+      `/data/productions/${obj.page.params["slug"]}.json`,
+    )
     const data = await res.json()
 
     if (res.status === 200) {
-      return { productions: data.productions, year: parseInt(params.slug) }
+      return {
+        props: {
+          productions: data.productions,
+          year: parseInt(obj.page.params["slug"]),
+        },
+      }
     } else {
-      this.error(res.status, data.message)
+      return { status: res.status, error: new Error(data.message) }
     }
   }
 </script>
