@@ -1,15 +1,18 @@
 <script lang="ts" context="module">
+  import type { Load } from "@sveltejs/kit"
+
   type Post = {
     content: string
     active: boolean
     title: string
   }
 
-  export function preload({ params, query }) {
-    return this.fetch(`jobs.json`)
+  export const load: Load = (obj) => {
+    return obj
+      .fetch(`/jobs.json`)
       .then((r) => r.json())
       .then((posts: Post[]) => {
-        return { posts: posts.filter((p) => p.active) }
+        return { props: { posts: posts.filter((p) => p.active) } }
       })
   }
 </script>
@@ -48,9 +51,9 @@
 
 <p class="bg-green-200 p-4 my-4">
   <strong>Notice:</strong>
-  Due to continuing health concerns, we are holding off on announcing our 2021
-  season until we have more information. We will update this page as soon as we
-  have finalized plans for 2021.
+  Due to continuing health concerns, we are holding off on announcing our 2021 season
+  until we have more information. We will update this page as soon as we have finalized
+  plans for 2021.
 </p>
 
 <div class="flex items-center my-4">
@@ -78,15 +81,15 @@
     <ul class="my-4 list-none p-0">
       <li>
         RSS:
-        <a
-          class="link-green"
-          href="{site.url}/jobs/feeds/rss">{site.url}/jobs/feeds/rss</a>
+        <a class="link-green" href="{site.url}/jobs/feeds/rss"
+          >{site.url}/jobs/feeds/rss</a
+        >
       </li>
       <li class="mt-2">
         JSON Feed:
-        <a
-          class="link-green"
-          href="{site.url}/jobs/feeds/json">{site.url}/jobs/feeds/json</a>
+        <a class="link-green" href="{site.url}/jobs/feeds/json"
+          >{site.url}/jobs/feeds/json</a
+        >
       </li>
     </ul>
 
