@@ -1,14 +1,16 @@
 <script lang="ts" context="module">
-  export async function preload({ params, query }) {
+  import type { Load } from "@sveltejs/kit"
+
+  export const load: Load = async (obj) => {
     // the `slug` parameter is available because
     // this file is called [slug].svelte
-    const res = await this.fetch(`data/productions.json`)
+    const res = await obj.fetch(`/data/productions.json`)
     const data = await res.json()
 
     if (res.status === 200) {
-      return { site: data.site, seasons: data.seasons }
+      return { props: { site: data.site, seasons: data.seasons } }
     } else {
-      this.error(res.status, data.message)
+      return { status: res.status, error: new Error(data.message) }
     }
   }
 </script>
