@@ -1,12 +1,20 @@
 <script lang="ts" context="module">
-  export async function preload({ params, query }) {
-    const res = await this.fetch(`data/people/${params.slug}.json`)
+  import type { Load } from "@sveltejs/kit"
+
+  export const load: Load = async (obj) => {
+    const res = await obj.fetch(`/data/people/${obj.page.params["slug"]}.json`)
     const data = await res.json()
 
     if (res.status === 200) {
-      return { site: data.site, people: data.people, slug: params.slug }
+      return {
+        props: {
+          site: data.site,
+          people: data.people,
+          slug: obj.page.params["slug"],
+        },
+      }
     } else {
-      this.error(res.status, data.message)
+      return { status: res.status, error: new Error(data.message) }
     }
   }
 </script>
@@ -30,20 +38,19 @@
     : "Cast, Musicians, Crew, and Staff"
 </script>
 
-{#each ['rest', 'Board', 'Additional'] as groupName}
-  {#if slug === '2020'}
+{#each ["rest", "Board", "Additional"] as groupName}
+  {#if slug === "2020"}
     <p>
       <a href="/news/2020-03-25-season-cancelled" class="link-green">
         Our 2020 Season was sadly cancelled.
       </a>
-      Preserved here are the people we were really looking forward to working
-      with.
+      Preserved here are the people we were really looking forward to working with.
     </p>
   {/if}
 
   {#if groupedPeople[groupName].length}
     <h2 class="h2 sm:sticky top-0 bg-white mb-4">
-      {groupName === 'rest' ? generalGroupName : groupName}
+      {groupName === "rest" ? generalGroupName : groupName}
     </h2>
     {#each groupedPeople[groupName] as person}
       <Bio person="{person}" />
